+++ forui/lib/widgets/button.dart
@@ -8,4 +8,3 @@ library forui.widgets.button;
 export '../src/widgets/button/button.dart' hide Variant;
 export '../src/widgets/button/button_content.dart' hide Content, IconContent;
 export '../src/widgets/button/button_styles.dart';
-export '../src/widgets/button/button_spinner.dart';